Maven

+++ Hotel1/repository-api/src/main/java/com/senlainc/git_courses/java_training/petushok_valiantsin/api/repository/IGuestDao.java
@@ -2,7 +2,7 @@ package com.senlainc.git_courses.java_training.petushok_valiantsin.api.repositor
 
 import com.senlainc.git_courses.java_training.petushok_valiantsin.model.Guest;
 
-public interface IGuestDao extends ICommonDao<Guest, Integer> {
+public interface IGuestDao extends ICommonDao<Guest, Long> {
 
-    Integer readSize();
+    Long readSize();
 }
