+++ Hotel1/repository-api/src/main/java/com/senlainc/git_courses/java_training/petushok_valiantsin/api/repository/IRoomDao.java
@@ -8,7 +8,7 @@ public interface IRoomDao extends ICommonDao<Room, Long> {
 
     List<Room> readAllFree();
 
-    Integer readFreeSize();
+    Long readFreeSize();
 
     Room readByNumber(Integer number);
 }
